refactor : Rq, RsDate를 활용하여 호감삭제 기능 수정

+++ src/main/java/com/ll/gramgram/boundedContext/likeablePerson/service/LikeablePersonService.java
@@ -1,5 +1,6 @@
 package com.ll.gramgram.boundedContext.likeablePerson.service;
 
+import com.ll.gramgram.base.rq.Rq;
 import com.ll.gramgram.base.rsData.RsData;
 import com.ll.gramgram.boundedContext.instaMember.entity.InstaMember;
 import com.ll.gramgram.boundedContext.instaMember.service.InstaMemberService;
@@ -7,8 +8,10 @@ import com.ll.gramgram.boundedContext.likeablePerson.entity.LikeablePerson;
 import com.ll.gramgram.boundedContext.likeablePerson.repository.LikeablePersonRepository;
 import com.ll.gramgram.boundedContext.member.entity.Member;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -51,8 +54,16 @@ public class LikeablePersonService {
     }
 
     @Transactional
-    public void deleteLikeablePerson(LikeablePerson likeablePerson) {
+    public RsData<LikeablePerson> deleteLikeablePerson(Integer id, InstaMember instaMember) {
+        LikeablePerson likeablePerson = getLikeablePerson(id);
+
+        if (instaMember.getId() != likeablePerson.getFromInstaMember().getId()) {
+            return RsData.of("F-1", "호감상대를 삭제할 권한이 없습니다.");
+        }
+
         likeablePersonRepository.delete(likeablePerson);
+
+        return RsData.of("S-1", "호감상대(%s)가 삭제되었습니다.".formatted(likeablePerson.getToInstaMemberUsername()), likeablePerson);
     }
 
     public LikeablePerson getLikeablePerson(Integer id) {
